% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add function is used to add one argument to another}
\usage{
add(x, y)
}
\arguments{
\item{x}{numeric, first parameter for add function}

\item{y}{numeric argument}
}
\description{
Add function is used to add one argument to another
}
