% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_of_square.R
\name{area_of_square}
\alias{area_of_square}
\title{!@title Square Area}
\usage{
area_of_square(length, width)
}
\arguments{
\item{_width}{the width of the square}

\item{_length}{the length of the square}
}
\description{
function calculates area of square from width and length
}
\author{
Runa Magnusson
}
