% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clustalRNA}
\alias{clustalRNA}
\title{alignment of RNA sequences for the spike protein of some sarbecoviruses
closely related to SARSCoV2}
\format{
a list with 4 elements \describe{ \item{nb}{number of sequences
compared} \item{nam}{sequence name (accession number)} \item{seq}{a list of
RNA sequences} \item{com}{NA} }
}
\source{
the \href{https://www.ncbi.nlm.nih.gov/genbank/}{genbank} database of
the National Center for Biotechnology Information or the Universal Protein
Resource \href{https://www.uniprot.org/}{UniProt} database
}
\usage{
clustalRNA
}
\description{
14 different sequences of some sarbecoviruses closely related to SARSCoV2
were retrieve from public databases and compared using CLUSTAL omega (1.2.4)
alignment software (see examples section for the complete protocol for
retrieving, saving and processing the data)
}
\keyword{datasets}
