% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countNA_total.R
\name{countNA_total}
\alias{countNA_total}
\title{Sum total NAs in dataset}
\usage{
countNA_total(dataframe)
}
\arguments{
\item{dataframe}{A dataset with NAs}

\item{data}{A dataset with NAs}
}
\value{
dataframe with total NAs
}
\description{
Sum total NAs in dataset
}
\examples{
dataset <- matrix(sample(c(NA, 1:5),25, replace = TRUE), 5)
df <- as.data.frame(dataset)
df
}
