% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discount.R
\name{discount}
\alias{discount}
\title{Two Discount functions}
\usage{
discount(listPrice, discountRate = 30/100)
}
\value{
product of listPrice abd discountRate
}
\description{
The two functions calculate the actual discount for 30% discount rate and custome discount rate
}
\details{
This first function simply multiply list price by 30% discount rate and the the second function multiplies list price by custom discount rate
}
\examples{
f(listPrice = 100, discountrate = 5/100)
}
\author{
Ibrahim Bello
}
