% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_cumprodtest.R
\name{f_cumprodtest}
\alias{f_cumprodtest}
\title{Cumulative product function that accounts for NAs}
\usage{
f_cumprodtest(x)
}
\arguments{
\item{x}{single numeric vector.}
}
\value{
A numeric vector of the same length as x.
}
\description{
The function \code{f_cumprodtest} calculates cumulative products of a numeric vector, adjusting for the occurence of NAs on left-hand side of sample.
}
\details{
This is merely a support function to deal with cumulative products of series that may or may not have NAs in the early part of their sample.
}
\examples{
require(zoo)
f_cumprodtest(c(NA, NA, 1.5, 2, 3))
f_cumprodtest(c(1.2, 1.5, 2, 3))
}

