% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{The map with the state ploted on the map

This is a function of creating map, the map plots the location of the record on the mao for the specif-
ied state and year.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer representing the index of the state}

\item{year}{An interger representing the specified year}
}
\value{
a map, it plots the location of the record on the map for the specified state
and year
}
\examples{
fars_map_state(state.num = 4, year = 2014)
fars_map_state(5, 2013)

}

