% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{Read .csv file

This is a function that, by default, reads .csv file.
If the file does not exist in the repository, the function
will stop. If the file exists, the function will return the
Data Frame Tbl of the .csv file.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the .csv file}
}
\value{
return the .csv file in the format of Data Frame Tbl
}
\examples{
\dontrun{
fars_read(filename = "accident_2013.csv")
fars_read("accident_2013.csv")
}

}

