% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{The count of records for the month of the specified year

This is a function returning the data frame of monthly record counts for the
specified year/years. The first column of the returned data frame is
the month, the second column is the record counts of the corresponding month
for the passed year/years.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{An integer year, a list of integer years and a vector of integer}
}
\value{
a data frame: the first column of the returned data frame is the
month. The rest of the column of the monthly counts for the passed year/years
}
\examples{
\dontrun{
fars_summarize_years(list(2013, 2014))
fars_summarize_years(c(2013, 2014))
fars_summarize_years(2013)
}

}

