% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fastaRNA}
\alias{fastaRNA}
\title{Production and farm value of maple products in Canada}
\format{
A data frame with columns:
\describe{
\item{Year}{A value between 1924 and 2015.}
\item{Syrup}{Maple products expressed as syrup, total in thousands of gallons.}
\item{CAD}{Gross value of maple products in thousands of Canadian dollars.}
\item{Region}{Postal code abbreviation for territory or province.}
}
\link{1} "QHD43416.1"     "AAP30030.1"     "QHR63300.2"     "AGZ48831.1"     "YP_001039971.1" "ALK02457.1"     "AVP78031.1"     "AVP78042.1"     "AAY88866.1"     "AGZ48818.1"
\link{11} "AAS00003.1"     "AAP51227.1"     "AAU04646.1"     "YP_003858584.1"
}
\source{
Statistics Canada. Table 001-0008 - Production and farm value of
maple products, annual. \url{http://www5.statcan.gc.ca/cansim/}
}
\usage{
fastaRNA
}
\description{
Production and farm value of maple products in Canada
}
\examples{
\dontrun{
maple
}
}
\keyword{datasets}
