% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstpackage-package.R
\docType{package}
\name{firstpackage-package}
\alias{firstpackage}
\alias{firstpackage-package}
\title{firstpackage: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/selesnow/firstpackage}
  \item \url{https://selesnow.github.io/firstpackage/}
  \item Report bugs at \url{https://github.com/selesnow/firstpackage/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
