% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_plus.R
\name{\%+\%}
\alias{\%+\%}
\title{Concatenate Strings}
\usage{
x \%+\% y
}
\arguments{
\item{x}{Character, first string}

\item{y}{Character, second string}
}
\value{
A character vector
}
\description{
Operator for concatenate two strings.
}
\details{
Use \verb{\%+\%} operator for join two or more strings
}
\section{Main}{

The main ......
}

\examples{
# two values
"first" \%+\% "second"

# three values
"first" \%+\% "second" \%+\% "three"
}
