% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{graph}
\alias{graph}
\title{Graph constructor
This function is for constructing graphs.}
\usage{
graph(nodes, links)
}
\arguments{
\item{nodes}{A vector of integers}

\item{links}{A list of vectors (edges)}
}
\value{
A list representing a graph.
}
\description{
Graph constructor
This function is for constructing graphs.
}
