% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/great.R
\name{greet}
\alias{greet}
\title{Greet a user}
\usage{
greet(name = "World")
}
\arguments{
\item{name}{character; name of user}
}
\value{
nothing! Just prints a greeting.
}
\description{
Greet a user
}
\examples{
#Default use
greet()

# supply user name
greet("Romy")
}
