% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inch2cm.R
\name{inch2cm}
\alias{inch2cm}
\title{Convert inch to cm}
\usage{
inch2cm(x)
}
\arguments{
\item{x}{a numeric vector containung values in inch}
}
\value{
a numeric vector containung values in cm
}
\description{
Convert inch to cm
}
\examples{
inch2cm(1:10)
}

