% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{intervals}
\alias{intervals}
\title{Compute intervals by week}
\arguments{
\item{dataset}{Data with 6 columns}

\item{by}{Compute fish passage by week}

\item{model}{Trap efficiency data}
}
\value{
Confidence intervals by week
}
\description{
Compute intervals by week
}
