% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelvin_to_celsius.R
\name{kelvin_to_celsius}
\alias{kelvin_to_celsius}
\title{Convert Kelvin to Celsius}
\usage{
kelvin_to_celsius(kelvin)
}
\arguments{
\item{kelvin}{Temperature in Kelvin scale.}
}
\value{
Temperature in Celsius scale.
}
\description{
Takes temperature value in Kelvin scales and returns value in Celsius scale.
}
\examples{
kelvin <- 310.15
kelvin_to_celsius(kelvin)

}
