% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot.R
\name{make_plot}
\alias{make_plot}
\title{make a line plot}
\usage{
make_plot(df, xlab, ylab)
}
\arguments{
\item{df}{name of the data frame}

\item{xlab}{name of x axis}

\item{ylab}{name of y axis}
}
\value{
a plot
}
\description{
make a line plot with ggplot
}
