% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{meanC}
\alias{meanC}
\title{Rcpp mean}
\usage{
meanC(x)
}
\arguments{
\item{x}{a numeric vector containung values}
}
\value{
numeric vector of length one containing the mean of \code{x}
}
\description{
using an Rcpp function to calculate the mean
}
\examples{
meanC(1:10)
}
\author{
Alexander Kowarik, \email{alexander.kowarik@statistik.gv.at}
}
