% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msfRNA}
\alias{msfRNA}
\title{Production and farm value of maple products in Canada}
\format{
A data frame with columns:
\describe{
\item{Year}{A value between 1924 and 2015.}
\item{Syrup}{Maple products expressed as syrup, total in thousands of gallons.}
\item{CAD}{Gross value of maple products in thousands of Canadian dollars.}
\item{Region}{Postal code abbreviation for territory or province.}
}
\link{1} "QHD43415_1"     "AAP30028_1"     "QHR63299_1"     "AGZ48830_1"     "YP_001039970_1" "ALK02468_1"     "AVP78030_1"     "AVP78041_1"     "AAY88865_2"     "AGZ48817_1"
\link{11} "AAS00002_1"     "AAP51225_1"     "AAU04645_1"     "YP_003858583_1"
}
\source{
Statistics Canada. Table 001-0008 - Production and farm value of
maple products, annual. \url{http://www5.statcan.gc.ca/cansim/}
}
\usage{
msfRNA
}
\description{
Production and farm value of maple products in Canada
}
\examples{
\dontrun{
maple
}
}
\keyword{datasets}
