% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K-nearest neighbors cross validation method}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{input data frame.}

\item{cl}{true class value of your training data.}

\item{k_nn}{integer representing the number of neighbors.}

\item{k_cv}{integer representing the number of folds.}
}
\value{
a list that contain objects;
*class: a vector of the predicted class $\hat{Y}_{i}$ for all observations
*cv_err: a numeric with the cross-validation misclassification error
}
\description{
This function predict output class \code{species} using covariates \code{bill_length_mm},
\code{bill_depth_mm}, \code{flipper_length_mm}, and \code{body_mass_g}. We will
be using a 5-fold cross validation to test whether it is working or not.
}
\examples{
my_knn_cv(data_p[, -1], data_p[, 1], 5, 5)

}
\keyword{cross}
\keyword{prediction,}
\keyword{validation.}
