% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Fitting Linear Models}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A "\code{formula}" class object that describes the model to
be fitted, typically having the form \code{response ~ terms}.}

\item{data}{Data frame containing the variables in the model.}
}
\value{
A table with rows for each coefficient and columns for the following:
  \code{Estimate}, \code{Std. Error}, \code{t value}, and \code{Pr(>|t|)}.
}
\description{
This function is used to fit linear models on a data set.
}
\examples{
# load mtcars data
data("mtcars")

my_lm(mpg ~ hp, mtcars)
my_lm(mpg ~ hp + wt, mtcars)

}
\keyword{inference}
