% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Integer of the number of folds.}
}
\value{
A numeric for the cross-validation error.
}
\description{
This function uses the \code{my_gapminder} data to predict \code{lifeExp}
using covariate \code{gdpPercap} using folds.
}
\examples{
my_rf_cv(5)

}
\keyword{prediction}
