% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t_test}
\alias{my_t_test}
\title{t-test}
\usage{
my_t_test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector of data values.}

\item{alternative}{Character string of the alternative hypothesis, must be
either "\code{two.sided}", "\code{less}", or "\code{greater}".}

\item{mu}{Numeric of true value of the mean.}
}
\value{
A list containing the following:

\code{t_stat}, the value of the test statistic.

\code{df}, the degrees of freedom for the t-statistic.

\code{alternative}, the alternative hypothesis.

\code{p_val}, the p-value for the test.
}
\description{
This function performs one-sample t-tests on vectors of data.
}
\examples{
my_t_test(1:10, "two.sided", 0)
my_t_test(rnorm(10), "less", 0)

}
\keyword{inference}
