% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myxbar.r
\name{MyMean}
\alias{MyMean}
\title{This function allows you to calculate the sample mean}
\usage{
MyMean(x)
}
\arguments{
\item{x}{the data}
}
\description{
This function allows you to calculate the sample mean
}
