% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one2many2.R
\name{one2many2}
\alias{one2many2}
\title{Deaggregate rows with rownames as integers instead of decimals}
\arguments{
\item{data}{A dataset with multiple rows}

\item{var}{The variable that the user want to split in multiple rows}
}
\value{
An aggregated dataset larger than the original
}
\description{
Deaggregate rows with rownames as integers instead of decimals
}
\examples{
p_load(tidyverse, Firstpackage)
a <- data.frame(name= c("Maria", "Tina"), age= c(5,3))
a
one2many2(a,'age')
}
