% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Reads data in 'file_name' as a data frame}
\usage{
read_data(file_name)
}
\arguments{
\item{file_name}{name of the file that contains bike counts}
}
\value{
a data frame with bike counts
}
\description{
read data file 'file_name' as a data fraame with external packages
}
