% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_error.R
\name{st_error}
\alias{st_error}
\title{Calculates standard error across a vector}
\usage{
st_error(vect)
}
\arguments{
\item{vect}{a vector of values for which the standard deviation can be computed}
}
\value{
numeric value of the standard error
}
\description{
Calculates standard error across a vector
}
