% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_lower_letters.R
\name{str_lower_letters}
\alias{str_lower_letters}
\title{converts upper letter to lower}
\usage{
str_lower_letters(napis)
}
\arguments{
\item{napis}{napis ktory chcemy zrobic wielkimi literami}
}
\value{
wyraz napisany malymi literami
}
\description{
converts upper letter to lower
}
\examples{
x <- 'EXAMPLE'
str_lower_letters(x)
}
