% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_upper_letters.R
\name{str_upper_letters}
\alias{str_upper_letters}
\title{Title}
\usage{
str_upper_letters(napis)
}
\arguments{
\item{napis}{napis ktory chcemy zrobic wielkimi literami}
}
\value{
napis zapisany wielkimi literami
}
\description{
Title
}
