% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_squares.R
\name{sum_squares}
\alias{sum_squares}
\title{Calculates the sum of squares for a sample}
\usage{
sum_squares(vect)
}
\arguments{
\item{vect}{a numeric vector over which to calculate the sum of squares}
}
\value{
Returns a number that is the calculated sum of squared for the given vector
}
\description{
Returns the sum of the squared deviations from mean, SS(x) of a numeric vector
}
