% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times2.R
\name{times2}
\alias{times2}
\title{Title function to double a number}
\usage{
times2(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
The nuber multiplied by 2
}
\description{
Title function to double a number
}
\examples{
times2(5)
}
