% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trainfile}
\alias{trainfile}
\title{DATASET OF ...............}
\format{
A data frame with ....... rows and .......... variables:
\describe{
\item{effort}{Effort measured in -------.}
\item{size}{Size measured in ....}
\item{source}{whatever indicating one source or another.}
}
}
\source{
\url{https://....domain.com... }
}
\usage{
trainfile
}
\description{
Data of effort and size for training
}
\details{
No missing values
A dataset containing -----  whatever you put here.
}
\keyword{datasets}
