\name{firstpackage-package}
\alias{firstpackage-package}
\alias{firstpackage}
\docType{package}
\title{
\packageTitle{firstpackage}
}
\description{
\packageDescription{firstpackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{firstpackage}
\packageIndices{firstpackage}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{firstpackage}

Maintainer: \packageMaintainer{firstpackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
