\name{firstPackage-package}
\alias{firstPackage-package}
\alias{firstPackage}
\docType{package}
\title{
  first Package
}
\description{
  Will contain functions to automate BLASTs
}
\details{
\tabular{ll}{
Package: \tab firstPackage\cr
Type: \tab Package\cr
Version: \tab 0.0.1-0\cr
Date: \tab 2012-02-21\cr
License: \tab GPL (>= 3)\cr
}
}
\author{
Author: Gerben Gaastra
Maintainer: Gerben Gaastra <gaastra@gtga.nl>
}
\references{
www.rug.nl/gbic
}
\keyword{ package }
\examples{
#TODO
}
