% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_plus.R
\name{\%+\%}
\alias{\%+\%}
\title{Concatinate}
\usage{
x \%+\% y
}
\arguments{
\item{x}{first string}

\item{y}{second string}
}
\value{
concatinate string
}
\description{
Concatinate
}
\examples{
"first" \%+\% "second"
}
