% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_tree}
\alias{make_tree}
\title{function to calculate a phylogenetic tree}
\usage{
make_tree(
  alignment,
  type,
  model = TN93,
  clustering = bionj,
  outgroup = NULL,
  plot = u
)
}
\arguments{
\item{alignment}{an object of class alignment or DNAbin containing a DNA, RNA or protein sequences alignment}

\item{type}{a character string without '' specifying the type of sequences, i.e. DNA, RNA or protein}

\item{model}{a character string without '' specifying the model to be used for the calculation of the distances matrix, i.e. raw, N, TS, TV, JC69, K80, F81, K81, F84, BH87, T92, TN93 (default), GG95, logdet, paralin, indel, or indelblock}

\item{clustering}{}

\item{outgroup}{}

\item{plot}{}
}
\value{
the function returns an object of class phylo of the ape package and a draft plot to visualize the phylogenetic calculations
}
\description{
this function is wrapper around the functions
}
\examples{
phylipProtTree <- makeTree(data(phylipProt), type=protein, model=K80, clustering=fastme.ols, outgroup=YP_0010399)
phylipRNATree <- makeTree(data(phylipRNA), type=RNA, clustering=fastme.bal, plot=clado)

}
