% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_upper_letters.R
\name{str_upper_letters}
\alias{str_upper_letters}
\title{convert the word from lower to upper letters}
\usage{
str_upper_letters(napis, test)
}
\arguments{
\item{napis}{napis ktory chcemy zrobic wielkimi literami}

\item{test}{czy on tez sie przepisze}
}
\value{
napis zapisany wielkimi literami
}
\description{
convert the word from lower to upper letters
}
\examples{
x <- 'example'
str_upper_letters(x)
}
